+++ CHANGELOG.md
@@ -9,13 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- "SPA action mode", the `SPA` option can now take a string, corresponding to a form action, and it will post there, without the need for a html form on the page.
+- "SPA action mode", the `SPA` option can now take a string, corresponding to a form action, and it will post there, even without a html form on the page.
 
 ### Fixed
 
 - Fixed types for constraints, tainted and errors when using intersections and unions in schemas.
 - Fixed SuperDebug collapsed height by preventing global css leak.
 - Redirect in SPA mode cancelled normal redirects, preventing applyAction.
+- Default objects sent to `superForm` returned a form with an empty id, causing collisions, it is now a random string.
 
 ## [2.8.1] - 2024-03-07